Adding more debug spew to figure out a permission issue
Adding more debug spew to figure out a permission issue.  Right now we are pulling down the package, decompressing it but on execution we get permission denied.   Adding some ls -al to dump out current folder and also scommand/bin.   Checking for permission type things.   Also increased the sleep time to 5 seconds before running.   Something odd.  Sucks I have to check in every time to test this but this is what we have.