Found issue with permission on decompression
Found issue with permission on decompression.  Before I fix this I want to dump out the error code on the system call mike has.   Because if we fail for whatever reason on this command we pull in the last result and that's wrong behavior.

So this change will just print that out before.   I am going to fix the permissions issue on the next change.